requirements.txt
click==6.7
fields==5.0.0
pkg-resources==0.0.0
PyVCF==0.6.8
#Tabby==0.4.0
-e git+https://github.com/signed0/tabby.git@master#egg=tabby
-e git+git@github.com:COMBAT-TB/tsv_validate_vcf.git@9c82c114c7a88e7df550eb0afea1daf5fcea0fe5#egg=vcf_snp_utils
